pythonBackend/requirements.txt
fastapi
uvicorn
pydantic-settings
httpx
tldextract
python-whois
pyopenssl
beautifulsoup4
aioftp
asyncssh